requirements.txt
colour==0.1.5
ezdxf==1.3.1
ifcopenshell==0.7.0.240627
laspy==2.5.4
lazrs
matplotlib==3.9.1
numpy_quaternion==2023.0.4
open3d
opencv_python
pandas==2.2.2
pillow==10.4.0
pye57==0.4.5
pyvista==0.44.0
rdflib==7.0.0
scikit_learn==1.5.1
scipy==1.14.0
trimesh==4.4.2
vtk==9.3.1
XlsxWriter==3.2.0
